requirements.txt
streamlit
tensorflow
numpy
Pillow
streamlit-option-menu
streamlit-webrtc
opencv-python-headless
